make transformer accept pd.Series; implement suggestions from aiwalter, add myself to list of contributors
